Fix the run of 5461dc048, whose unit-test-coverage-shard-17 found ScopeRandomizeWith.IntArgumentsRandomizedAgainstStateValue of test/src/unit/test_simulator_subclause_18_12_01a.cpp reading a drawn sum of two task-local ints, a and b below 5 and their sum below a length of 6, as 3639850998 against its bound of 6: the case was written while an int of a task was created unsigned, when no draw was negative, and it compared the sum's bit pattern as an unsigned value. c2f9e0d2c gave a task's int its declared signedness, so a below 5 now admits a negative a and the drawn sum of that run was the negative int §6.11.2 declares it, valid under every relation of the call; the case now reads the sum as the int it is, and what it states, that two int arguments are drawn jointly against an int state variable, is unchanged
Refs #1733

+++ test/src/unit/test_simulator_subclause_18_12_01a.cpp
@@ -210,8 +210,11 @@ TEST(ScopeRandomizeWith, IntArgumentsRandomizedAgainstStateValue) {
       "  endtask\n"
       "  initial run(6);\n"
       "endmodule\n";
-  EXPECT_EQ(RunAndGet(src, "okv"), 1u);   // a joint int solution below 6 exists
-  EXPECT_LT(RunAndGet(src, "sumv"), 6u);  // and the drawn int sum respects it
+  EXPECT_EQ(RunAndGet(src, "okv"), 1u);  // a joint int solution below 6 exists
+  // The sum is a signed int (§6.11.2), and a below 5 admits a negative a, so
+  // the sum is read as the int it is: its bit pattern as an unsigned value
+  // would rank a negative sum above 6.
+  EXPECT_LT(static_cast<int32_t>(RunAndGet(src, "sumv")), 6);
 }
 
 // 18.12.1: a non-argument operand a with-block reads as a state constant may be
